feat: demote the pipeline pod restart alert to a warning
This is related to https://github.com/redhat-appstudio/o11y/pull/411

In that PR, we added a new alert that fires when pods are in
crashloopbackoff for more than 3 minutes. If they enter that state,
something is definitely down.

The alert being modified in this change fires whenever pods restart more
than once in a short period. Today, I observed this fire when the remote
resolver controller pods all restarted a few times at the same time (4
pods, 2-3 restart each). They immediately recovered and began working
happily. No perceived outage for users as far as I can tell.

If I understand that situation correctly, it means that this isn't
really an SLO. It's more like a warning that can let us know if things
are about to enter sustained crashloop backoff.

Signed-off-by: Ralph Bean <rbean@redhat.com>